backend/requirements.txt
openai==0.27.9
python-dotenv==1.0.0
